edit pass: mysql-database-enus-business-continuity-disaster-recovery.md
H2 headings were numbered. I removed the numbers because the sections do not represent a sequence.
All content in two notes was bolded, which is not our style. It's customary to bold selective words, but it's not our style to bold all text in a note.
The "Common service continuity problems" section included a promise of up upcoming feature. I deleted this text because the guidance is to avoid promises of upcoming features. When a feature is available, content can mention it.